Installed Wirble.
Wirble is a Ruby gem that makes `irb' a little nicer and easier to use.
In particular, it adds tab-completion, history, and a built-in `ri'
command.  I do not use colorization feature.  Tab-completion doesn't
work in inferior Ruby buffer in Emacs (which is one of the major
drawbacks of all inferior interpreter modes in Emacs), but `ri' command
does work.  I installed Wirble locally, hence exporting GEM_HOME in
.profile.  Also, we have to set this variable manually in .emacs.